tests: repair tests to use Double

+++ app/src/test/java/com/example/wellfed/StorageIngredientTest.java
@@ -36,11 +36,11 @@ public class StorageIngredientTest {
      */
     @Test
     public void testConstructor2() {
-        StorageIngredient ingredient = new StorageIngredient("Apple", 1.0f,
+        StorageIngredient ingredient = new StorageIngredient("Apple", 1.0,
                 "g",
                 "Pantry", new Date(2020, 1, 1));
         assertEquals("Apple", ingredient.getDescription());
-        assertEquals((Float) 1.0f, ingredient.getAmount());
+        assertEquals((Double) 1.0, ingredient.getAmount());
         assertEquals("g", ingredient.getUnit());
         assertEquals("Pantry", ingredient.getLocation());
         assertEquals(new Date(2020, 1, 1), ingredient.getBestBeforeDate());
@@ -58,11 +58,11 @@ public class StorageIngredientTest {
      */
     @Test
     public void testConstructor3() {
-        StorageIngredient ingredient = new StorageIngredient("Apple", 1.0f,
+        StorageIngredient ingredient = new StorageIngredient("Apple", 1.0,
                 "g",
                 "Pantry", new Date(2020, 1, 1), "Fruit");
         assertEquals("Apple", ingredient.getDescription());
-        assertEquals((Float) 1.0f, ingredient.getAmount());
+        assertEquals((Double) 1.0, ingredient.getAmount());
         assertEquals("g", ingredient.getUnit());
         assertEquals("Pantry", ingredient.getLocation());
         assertEquals(new Date(2020, 1, 1), ingredient.getBestBeforeDate());
@@ -88,11 +88,11 @@ public class StorageIngredientTest {
         assertNull(ingredient.getBestBefore());
         assertNull(ingredient.getCategory());
 
-        StorageIngredient ingredient2 = new StorageIngredient("Apple", 1.0f,
+        StorageIngredient ingredient2 = new StorageIngredient("Apple", 1.0,
                 "g",
                 "Pantry", new Date(2020, 1, 1));
         assertEquals("Apple", ingredient2.getDescription());
-        assertEquals((Float) 1.0f, ingredient2.getAmount());
+        assertEquals((Double) 1.0, ingredient2.getAmount());
         assertEquals("g", ingredient2.getUnit());
         assertEquals("Pantry", ingredient2.getLocation());
         assertEquals(new Date(2020, 1, 1), ingredient2.getBestBeforeDate());
@@ -104,11 +104,11 @@ public class StorageIngredientTest {
         assertEquals(bestBeforeString, ingredient2.getBestBefore());
         assertNull(ingredient2.getCategory());
 
-        StorageIngredient ingredient3 = new StorageIngredient("Apple", 1.0f,
+        StorageIngredient ingredient3 = new StorageIngredient("Apple", 1.0,
                 "g",
                 "Pantry", new Date(2020, 1, 1), "Fruit");
         assertEquals("Apple", ingredient3.getDescription());
-        assertEquals((Float) 1.0f, ingredient3.getAmount());
+        assertEquals((Double) 1.0, ingredient3.getAmount());
         assertEquals("g", ingredient3.getUnit());
         assertEquals("Pantry", ingredient3.getLocation());
 
@@ -130,8 +130,8 @@ public class StorageIngredientTest {
         StorageIngredient ingredient = new StorageIngredient("Apple");
         ingredient.setDescription("Banana");
         assertEquals("Banana", ingredient.getDescription());
-        ingredient.setAmount(1.0f);
-        assertEquals((Float) 1.0f, ingredient.getAmount());
+        ingredient.setAmount(1.0);
+        assertEquals((Double) 1.0, ingredient.getAmount());
         ingredient.setUnit("g");
         assertEquals("g", ingredient.getUnit());
         ingredient.setLocation("Pantry");
@@ -147,13 +147,13 @@ public class StorageIngredientTest {
         ingredient.setCategory("Fruit");
         assertEquals("Fruit", ingredient.getCategory());
 
-        StorageIngredient ingredient2 = new StorageIngredient("Apple", 1.0f,
+        StorageIngredient ingredient2 = new StorageIngredient("Apple", 1.0,
                 "g",
                 "Pantry", new Date(2020, 1, 1));
         ingredient2.setDescription("Banana");
         assertEquals("Banana", ingredient2.getDescription());
-        ingredient2.setAmount(1.0f);
-        assertEquals((Float) 1.0f, ingredient2.getAmount());
+        ingredient2.setAmount(1.0);
+        assertEquals((Double) 1.0, ingredient2.getAmount());
         ingredient2.setUnit("g");
         assertEquals("g", ingredient2.getUnit());
         ingredient2.setLocation("Pantry");
@@ -169,13 +169,13 @@ public class StorageIngredientTest {
         ingredient2.setCategory("Fruit");
         assertEquals("Fruit", ingredient2.getCategory());
 
-        StorageIngredient ingredient3 = new StorageIngredient("Apple", 1.0f,
+        StorageIngredient ingredient3 = new StorageIngredient("Apple", 1.0,
                 "g",
                 "Pantry", new Date(2020, 1, 1), "Fruit");
         ingredient3.setDescription("Banana");
         assertEquals("Banana", ingredient3.getDescription());
-        ingredient3.setAmount(1.0f);
-        assertEquals((Float) 1.0f, ingredient3.getAmount());
+        ingredient3.setAmount(1.0);
+        assertEquals((Double) 1.0, ingredient3.getAmount());
         ingredient3.setUnit("g");
         assertEquals("g", ingredient3.getUnit());
         ingredient3.setLocation("Pantry");
@@ -200,20 +200,20 @@ public class StorageIngredientTest {
         StorageIngredient ingredient = new StorageIngredient("Apple");
         StorageIngredient ingredient2 = new StorageIngredient("Apple");
         StorageIngredient ingredient3 = new StorageIngredient("Banana");
-        StorageIngredient ingredient4 = new StorageIngredient("Apple", 1.0f,
+        StorageIngredient ingredient4 = new StorageIngredient("Apple", 1.0,
                 "g",
                 "Pantry", new Date(2020, 1, 1));
-        StorageIngredient ingredient5 = new StorageIngredient("Apple", 1.0f,
+        StorageIngredient ingredient5 = new StorageIngredient("Apple", 1.0,
                 "g",
                 "Pantry", new Date(2020, 1, 1));
-        StorageIngredient ingredient6 = new StorageIngredient("Apple", 1.0f,
+        StorageIngredient ingredient6 = new StorageIngredient("Apple", 1.0,
                 "g",
                 "Pantry", new Date(2020, 1, 2));
-        StorageIngredient ingredient7 = new StorageIngredient("Test4", 1.0f,
+        StorageIngredient ingredient7 = new StorageIngredient("Test4", 1.0,
                 "g", "Pantry", new Date(2020, 1, 1), "Fruit");
-        StorageIngredient ingredient8 = new StorageIngredient("Test5", 1.0f,
+        StorageIngredient ingredient8 = new StorageIngredient("Test5", 1.0,
                 "g", "Pantry", new Date(2020, 1, 1), "Fruit");
-        StorageIngredient ingredient9 = new StorageIngredient("Test7", 1.0f,
+        StorageIngredient ingredient9 = new StorageIngredient("Test7", 1.0,
                 "g", "Pantry", new Date(2020, 1, 1), "Fruit2");
         assertTrue(ingredient.equals(ingredient2));
         assertFalse(ingredient.equals(ingredient3));
